# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parkly API Documentation",
        "description": "API documentation for Parkly backend",
        "contact": {
            "email": "stefan.pejchinoski@iwconnect.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://16.171.224.141",
            "description": "API Server"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/parking-spots": {
            "get": {
                "tags": [
                    "Parking Spots"
                ],
                "summary": "Get a list of all parking spots",
                "description": "Display a listing of parking spots.",
                "operationId": "20ad82c7df831eb568326203b804a995",
                "parameters": [
                    {
                        "name": "facility_id",
                        "in": "query",
                        "description": "Filter spots by facility ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "facility_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "spot_number": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "facility": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Downtown Parking"
                                                            },
                                                            "parking_spot_count": {
                                                                "type": "integer",
                                                                "example": 50
                                                            },
                                                            "manager_id": {
                                                                "type": "integer",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parking Spots"
                ],
                "summary": "Create a new parking spot",
                "description": "Store a newly created parking spot.",
                "operationId": "5c6d681ecd82664aa22fdc104e637d68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "facility_id",
                                    "spot_number"
                                ],
                                "properties": {
                                    "facility_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "spot_number": {
                                        "type": "integer",
                                        "example": 6
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking spot created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "facility_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "spot_number": {
                                                    "type": "integer",
                                                    "example": 6
                                                },
                                                "facility": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Downtown Parking"
                                                        },
                                                        "parking_spot_count": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "manager_id": {
                                                            "type": "integer",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "spot_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The spot number is already taken for this facility."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - User doesn't have permission"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}